Fix wrong publishing procedure and setup
Due to a misinterpretation the previous publishing procedure (as well as the established) was not optimal in any sense, not to mention that it was partially incorrect. This has the direct consequence of the malformation of the RSS files.

It is currently possible to work in production and development mode without any of the modes interfering with each other (this was previously not possible).